% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calageInt.R
\name{calageInt}
\alias{calageInt}
\title{Age Intensity Function}
\usage{
calageInt(state, proxel, delta)
}
\arguments{
\item{state}{a string value}

\item{proxel}{a data frame containing a state, age intensity and the probability.}

\item{delta}{a numeric value as time step}
}
\value{
a numeric value as the age intensity
}
\description{
This function returns a numeric value indicating
the pending time (in terms of time steps, delta) for a given state change of a proxel.
}
\examples{
proxel <- data.frame(State = "OK", ageInt = 0, Prob = 1)
state <- "OK"
delta <- 0.2
calageInt(state, proxel,delta)
}
