% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calProb.R
\name{calProb}
\alias{calProb}
\title{Transition Probability Function}
\usage{
calProb(BE, state, proxel, delta)
}
\arguments{
\item{BE}{a list containing states, transition matrix, distributions and their parameters for a basic event}

\item{state}{a string value}

\item{proxel}{a data frame containing a state, age intensity and a probability.}

\item{delta}{a numeric value as time step}
}
\value{
a numeric value between 0 and 1 as the transition probability
}
\description{
This function returns the transition probability for a given basic event and a proxel.
}
\examples{
## A repairable basic event with Uniform(2, 2.5) failure distribution function
## and a fixed repair time of 0.3.
delta <- 0.2
BE <- list(
  states = c("OK", "F"),
  G = rbind(
    c(NA, 1),
    c(1, NA)
  ),
  dist = c("unif", "unif"),
  param = list(c(2, 2.5), c(0.3 - delta, 0.3 + delta))
)
state <- "OK"
proxel <- data.frame(State = "OK", ageInt = 0, Prob = 1)
calProb(BE, state, proxel, delta)
}
