% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{h}
\alias{h}
\title{Hazard Rate Function}
\usage{
h(t, delta, D, P, ...)
}
\arguments{
\item{t}{a numeric value as time}

\item{delta}{a numeric value  time step}

\item{D}{density function}

\item{P}{cumulative density function}
}
\value{
A numeric vector of hazard rate values.
}
\description{
This function calculates ....
}
\details{
The more details should be added.
}
\examples{
t <- 0.1
delta <- 0.01
P <- pnorm
D <- dnorm
h(t, delta, D,P)
}
