% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEUna.R
\name{TEUna}
\alias{TEUna}
\title{The title of function}
\usage{
TEUna(mcs, una)
}
\arguments{
\item{mcs}{a ...}

\item{una}{a ...}
}
\value{
A numeric matrix of transition probabilities.
}
\description{
This function returns a ....
}
\details{
The more details should be added.
}
\examples{
...

}
